Remove warning for 'implicit declaration of mkdir'
Before this, there would be a warning of 'implicit declaration of
mkdir' for the miniunz.c file on line 131 where it calls `mkdir`.
That is bad because there is a warning.

The fix is to include the appropriate header files as indicated
by `man(2) mkdir`; `sys/stat.h` and `sys/types.h`.

This fix is for unix flavoured systems, but may need to be tweaked
for some flavours.