Fix warning for lines 521 & 529 in minizip/zip.c
Before this, there would be a warning for the lines 521 and 529
in contrib/minizip/zip.c. This was caused by the `for` loop on
line 521 not having braces around the body of the `for`, and that
the subsequent statement on line 529 was indented to look like it
was part of the `for` body.

This is now corrected by fixing the bad indentation.

The alternative of placing braces around the `for` body was considered
but that style is not typically used in the zlib codebase.